pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "arches_rascolls"
readme = "README.md"
authors = []
license = {text = "GNU AGPL3"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
]
requires-python = ">=3.11"
dependencies = [
    "arches>=8.1.0a0,<8.2.0",
    "django_csp==3.8",
    "arches_modular_reports @ git+https://github.com/archesproject/arches-modular-reports@main",
    "arches_querysets @ git+https://github.com/archesproject/arches-querysets@main",
    "arches_controlled_lists @ git+https://github.com/archesproject/arches-controlled-lists@main",
    "arches_search @ git+https://github.com/archesproject/arches-search@main",
]
version = "0.0.1"

[dependency-groups]
dev = [
    "livereload",
    "sst",
    "coverage",
    "django-silk==5.1.0",
    "pre-commit==3.8.0",
    "black==24.4.2",
]

[tool.setuptools]
packages = ["arches_rascolls"]
